Implement initial state in index.js
